logout vulnerability
when a user logs out, the constant containing the token still has it due to persistence, so ... as a temporary fix, simply reload the page everytime the user logsout. As, reloading by default doesn't delete the token but it ensures to reload all the webapp resources, here the admintoken. (This might not be the best fix yet, might shift to another way to fix this is something pops us).